Thu, Aug 4, 2022, 9:40 AM +00:00

+++ week2/readability/readability.c
@@ -30,7 +30,7 @@ int main(void)
     }
     else
     {
-        printf("Grade %i\n", round(index));
+        printf("Grade %i\n", (int) round(index));
     }
 }
